docs: updated diagrams
The diagrams correspond to the current status of the implementation. Omitted are the details about the assembly trampoline and from where exactly it is loaded to the memory by the Linux kernel. At this point of the time I am exploring possibilities of doing that, and hence the final decision is not yet made.